Move lighten and darken method to dedicated ColorEditor class
