+++ mobile/src/androidTest/java/com/novoda/canvas/base/RandomColorFactory.java
@@ -22,23 +22,4 @@ public class RandomColorFactory {
         int blue = BASE_LEVEL + random.nextInt(TOP_UP_LEVEL);
         return Color.rgb(red, green, blue);
     }
-
-    @ColorInt
-    public int updateBrightness(@ColorInt int color, int delta) {
-        int red = updateComponentBrightness(Color.red(color), delta);
-        int green = updateComponentBrightness(Color.green(color), delta);
-        int blue = updateComponentBrightness(Color.blue(color), delta);
-        return Color.rgb(red, green, blue);
-    }
-
-    private int updateComponentBrightness(int component, int delta) {
-        int lightenValue = component + delta;
-        if (lightenValue > 255) {
-            lightenValue = 255;
-        }
-        if (lightenValue < 0) {
-            lightenValue = 0;
-        }
-        return lightenValue;
-    }
 }
